+++ packages/template-retail-react-app/app/pages/account/profile.jsx
@@ -100,15 +100,8 @@ const ProfileCard = ({allowPasswordChange = false}) => {
                     body: {
                         firstName: values.firstName,
                         lastName: values.lastName,
-                        phoneHome: values.phone,
-                        // NOTE/ISSUE
-                        // The sdk is allowing you to change your email to an already-existing email.
-                        // I would expect an error. We also want to keep the email and login the same
-                        // for the customer, but the sdk isn't changing the login when we submit an
-                        // updated email. This will lead to issues where you change your email but end
-                        // up not being able to login since 'login' will no longer match the email.
-                        email: values.email,
-                        login: values.email
+                        phoneHome: values.phone
+                        // Email field is now readonly and not included in the update
                     }
                 },
                 {